clarify N_ab is across all messages using one key
Clarify that N_ab is not a byte count for just one message or one conversation, but a total across all messages used with a key. This is to confirm message synchronization and more importantly to be sure we re-key frequently enough for to defeat AES-GCM cryptanalysis.

Thanks to Adam Wick for the clarification.