hotline networking: rename ROCCS as Broker
Alexa and Leah clarified for me that CATALINK uses the term ROCCS to refer to the resilient mesh network, not the particular proxy device that is the focus of this doc. I've renamed that device as "Broker" here, to convey the role of middleman, and also that it acts on behalf one one principal while dealing with other principals. In the back of my mind, it evokes "honest broker."