fill in initial gaps
After filling in the initial gaps, this code is working well enough for an initial demo inside acme. While the ROCCS is being built, I run a standalone broker that just copies message files between directories, accessed by individual's pucks using puckfs. That test infrastructure is not essential to understanding, so not copying it here yet.