hotline cryptography.pdf: puck command line
Mostly adding a cheat-sheet of "hotline fetch; hotline send"  and "hotline rekey" but also improve the text in a few places.

The only substantive change is to switch from JPEG to fax as the image format. We haven't implemented it yet, though should be easy using the Go semi-standard library x/image/ccitt.