puckfs: simplify protocol between puck and broker
I had thought the Ack/QAck messages would simplify the code, but it only made a mess. Revert to previous design. This mostly changes the wire protocol "ClientRPC" part of puckfs.go. Since this wire protocol is not compatible with the one in the few existing pucks that have been distributed, to reduce confusion I updated go.mod to call the executable "hotline2".

Our assumed production puck/broker network is a locally controlled link of low loss and low latency, so it seems ok to rely on timeouts. If using this over the general internet, probably one should just use TCP. That leads to lots of code complexity, but would only be for some prototype anyway.

I took the opportunity to clean up wording here and there, added interrupt signal handling for the broker, and tweaked the recently introduced "make clean" less catastrophic in the hands of a newcomer.