initial upload of version 3, primarily changing ciphers
Change the crypto to xchacha20poly1305, so that we can avoid the
need to rekey after a modest number of messages.

There is also some effort to make it easier to review the code, but
of course that is a matter of taste. If anyone really does look at this
closely, I'd be keen to hear.

This version does not have much testing yet.