puckfs/serverAPI.go: silence "value never used" warning
    file, req, err = extractFilename(req)
adopts the common idiom of peeling off part of the input and
returning the remainder. This is great, but the problem is in some
cases the remainder matters and some cases not. Worse, as people
change the code later, if you use _ in the assignment, it is easy to
overlook adding back req. So I made the more unused annotation
explicit with
    _ = req
to get rid of the warning from staticcheck.