remove stale reference to obsolete cipher
We have no specific knowledge of a problem with ascon, but since even the authors of that cipher recommend something new for post-quantum-crypto, we've dropped the reference to it.

The original intent in using that was to bend over backwards to avoid a US-centric cryptographic design. But each country decides which national champion cipher to use anyway, so it doesn't really matter.